push

+++ store/rootmulti/store.go
@@ -840,7 +840,7 @@ loop:
 			}
 
 		case *snapshottypes.SnapshotItem_AppVersion:
-			err := rs.SetAppVersion(uint64(item.AppVersion.Version))
+			err := rs.SetAppVersion(item.AppVersion.Version)
 			if err != nil {
 				return snapshottypes.SnapshotItem{}, sdkerrors.Wrap(err, "IAVL node import failed - error saving app version from received snapshot")
 			}
